Rename write_oz to write_provider
