+++ src/ethereum/write_provider/error.rs
